requirements.txt
Django
Pillow
djangorestframework
dj-rest-auth
django-allauth